requirements.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.4.2
Django==3.1.7
django-comments-dab==2.6.0
django-crispy-forms==1.11.2
django-mathfilters==1.0.0
django-taggit==1.3.0
gunicorn==20.1.0
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
